Rename tailwind config file
When I ran the `build` or `develop` commands on my Windows machine, Windows saw the `tailwind` program name and attempted to execute `tailwind.js` with its native JS engine (not Node).

The most straight-forward fix I could think of was to rename the config file, to avoid confusion between the command and the file. I've changed it to `tailwind.config.js` (a name I found commonly used elsewhere). With this change, the command works on my machine, now.

I also refactored the tailwind command into its own `build:css` script, to avoid having to change it in two places.